refactor(DetailNotice): 공고 상세 페이지 컴포넌트 분리
